feat: add UI trigger for quota snapshot migration on Quotas page
