+++ packages/frontend/src/lib/api.ts
@@ -2268,6 +2268,34 @@ export const api = {
     }
   },
 
+  getLegacySnapshotStatus: async (): Promise<{ tableExists: boolean; rowCount: number } | null> => {
+    try {
+      const res = await fetchWithAuth(`${API_BASE}/v0/management/quotas/legacy-snapshot-status`);
+      if (!res.ok) throw new Error('Failed to get legacy snapshot status');
+      return (await res.json()) as { tableExists: boolean; rowCount: number };
+    } catch (e) {
+      console.error('API Error getLegacySnapshotStatus', e);
+      return null;
+    }
+  },
+
+  migrateLegacySnapshots: async (): Promise<{
+    inserted: number;
+    skipped: number;
+    totalSource: number;
+  } | null> => {
+    try {
+      const res = await fetchWithAuth(`${API_BASE}/v0/management/quotas/migrate-legacy-snapshots`, {
+        method: 'POST',
+      });
+      if (!res.ok) throw new Error('Failed to migrate legacy snapshots');
+      return (await res.json()) as { inserted: number; skipped: number; totalSource: number };
+    } catch (e) {
+      console.error('API Error migrateLegacySnapshots', e);
+      return null;
+    }
+  },
+
   deleteAlias: async (aliasId: string): Promise<void> => {
     const res = await fetchWithAuth(
       `${API_BASE}/v0/management/models/${encodeURIComponent(aliasId)}`,
